Made CycleCount in EDRV_CYCLIC_USE_DIAGNOSTICS unsigned long long again

+++ Include/edrv.h
@@ -205,7 +205,7 @@ typedef struct
 typedef struct
 {
     // continuous min/max/avg measurement
-    unsigned long       m_ullCycleCount;
+    unsigned long long  m_ullCycleCount;
     DWORD               m_dwCycleTimeMin;
     DWORD               m_dwCycleTimeMax;
     unsigned long long  m_ullCycleTimeMeanSum;  // sums run over after some years for ct=400
